Enum as proc
Enums can be dynamic like aasm_state or items in database, so if representer have documentation with calls to db it will affect speed of representer. It is better to have proc to defer reading enum property.

Now that I think about this. It will be better to documentation property as proc. What do you think?